requirements/base.txt
Django==4.0.3
djangorestframework==3.13.1
psycopg2-binary==2.9.3
Pillow==9.0.1
drf-yasg==1.20.0
flex==6.14.1
swagger-spec-validator==2.7.4
django-dotenv==1.4.2
django-filter==21.1
django-extra-fields==3.0.2
djangorestframework-jwt==1.11.0
django-environ==0.8.1
gunicorn==20.1.0
requests==2.27.1
future==0.18.2
django-cors-headers==3.11.0
django-model-utils==4.2.0
jsonschema==3.2.0
redis==4.2.0
django-redis==5.2.0
celery==5.2.3
flower==1.0.0
whitenoise==6.4.0